alignment of directory structure
