--- App/__init__.py
@@ -1,24 +1,0 @@
-from flask import Flask
-from flask_sqlalchemy import SQLAlchemy
-from flask_migrate import Migrate
-
-db = SQLAlchemy()
-migrate = Migrate()
-
-def create_app(config_name):
-    app = Flask(__name__)
-    app.config.from_object(config_name)
-
-    db.init_app(app)
-    migrate.init_app(app, db)
-
-    # Registration of Blueprints
-    from .routes.auth_routes import auth_bp
-    from .routes.inventory_routes import inventory_bp
-    from .routes.report_routes import report_bp
-
-    app.register_blueprint(auth_bp)
-    app.register_blueprint(inventory_bp)
-    app.register_blueprint(report_bp)
-
-    return app
