--- App/routes/auth_routes.py
@@ -1,7 +1,0 @@
-from flask import Blueprint, request, jsonify
-from ..models import User, Invitation
-from .. import db
-
-auth_bp = Blueprint('auth', __name__)
-
-# authentication routes can be defined here
