+++ server/app/models.py
@@ -1,0 +1,21 @@
+from . import db
+from datetime import datetime
+
+class Inventory(db.Model):
+    __tablename__ = 'inventory'
+    inventory_id = db.Column(db.Integer, primary_key=True)
+    product_id = db.Column(db.Integer, db.ForeignKey('products.product_id'), nullable=False)
+    store_id = db.Column(db.Integer, db.ForeignKey('stores.store_id'), nullable=False)
+    quantity_received = db.Column(db.Integer, nullable=False)
+    quantity_in_stock = db.Column(db.Integer, nullable=False)
+    quantity_spoilt = db.Column(db.Integer, nullable=False)
+    payment_status = db.Column(db.String(10), nullable=False)
+    created_at = db.Column(db.DateTime, default=datetime.utcnow)
+
+class SupplyRequest(db.Model):
+    __tablename__ = 'supply_requests'
+    request_id = db.Column(db.Integer, primary_key=True)
+    inventory_id = db.Column(db.Integer, db.ForeignKey('inventory.inventory_id'), nullable=False)
+    user_id = db.Column(db.Integer, db.ForeignKey('users.user_id'), nullable=False)
+    request_date = db.Column(db.DateTime, default=datetime.utcnow)
+    status = db.Column(db.String(10), default='pending')
